[DOCS] Have examples use int for count

+++ tests/internal/test_compile.py
@@ -46,7 +46,7 @@ FOREVER = PipeDef(
                 {
                     "id": "sw-2",
                     "type": "truncate",
-                    "conf": TruncateRawConf({"count": {"type": "float", "value": "2"}}),
+                    "conf": TruncateRawConf({"count": {"type": "int", "value": "2"}}),
                 }
             ),
             PipeModule({"id": "_OUTPUT", "type": "output", "conf": {}}),
@@ -258,7 +258,7 @@ def test_convert_dag_linear_default_matches_explicit_wires():
             {
                 "id": "sw-2",
                 "type": "truncate",
-                "conf": {"count": {"type": "float", "value": "3"}},
+                "conf": {"count": {"type": "int", "value": "3"}},
             }
         ),
     ]
@@ -284,7 +284,7 @@ def test_convert_dag_generates_ids_when_omitted():
                 DagModule(
                     {
                         "type": "truncate",
-                        "conf": {"count": {"type": "float", "value": "3"}},
+                        "conf": {"count": {"type": "int", "value": "3"}},
                     }
                 ),
             ]
